@@ -69,8 +69,13 @@ class _MyHomePageState extends State<MyHomePage> {
                   });
                 },
                 decoration: InputDecoration(
-                  labelText: 'Start typing to validate...',
-                    suffixIcon: brandIcon != null ? FaIcon(brandIcon, size: 32,) : null),
+                    labelText: 'Start typing to validate...',
+                    suffixIcon: brandIcon != null
+                        ? FaIcon(
+                            brandIcon,
+                            size: 32,
+                          )
+                        : null),
               ),
             ),
             SizedBox(
